ml2/requirements.txt
Flask
python-dotenv
langchain-google-genai
flask-cors